requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
boolean.py==5.0
CacheControl==0.14.3
cachetools==4.2.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
cryptography==45.0.5
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dparse==0.6.4
filelock==3.16.1
fiona==1.10.1
frozenlist==1.7.0
geographiclib==2.0
geopandas==1.1.1
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
license-expression==30.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
nltk==3.9.1
numpy==2.3.1
packageurl-python==0.17.1
packaging==25.0
pandas==2.3.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
propcache==0.3.2
psutil==6.1.1
py-serializable==2.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.7
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rtree==1.4.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safety==3.6.0
safety-schemas==0.0.14
schedule==1.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
